#46 #48 Finalize logic for vat and shipping costs caluculation and

+++ db/migrate/20161030122204_create_taxes.rb
@@ -2,7 +2,7 @@ class CreateTaxes < ActiveRecord::Migration[5.0]
   def change
     create_table :taxes do |t|
       t.string :country
-      t.integer :vat_rate
+      t.decimal :vat_rate
 
       t.timestamps
     end
